102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i;
-	unsigned long long int fib1 = 1, fib2 = 2, fib_sum;
+	unsigned long fib1 = 1, fib2 = 2, fib_sum;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (i = 2; i < 50; i++)
 	{
 		fib_sum = fib1 + fib2;
-		printf(", %llu", fib_sum);
+		printf(", %lu", fib_sum);
 		fib1 = fib2;
 		fib2 = fib_sum;
 	}
